\name{aggregatedDat}
\alias{aggregatedDat}
\docType{data}
\title{Aggregated data}
\description{Aggregated data set, used for examples only}
\usage{data(aggregatedDat)}
\format{A data frame consisting of 4 grouping variables, the according Frequencies and the values of a numerical variable.}
\keyword{datasets}