% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{create_node}
\alias{create_node}
\alias{add_nodes}
\alias{delete_nodes}
\alias{rename_node}
\title{Create a hierarchy}
\usage{
create_node(...)

add_nodes(...)

delete_nodes(...)

rename_node(...)
}
\description{
create_node() is defunct, please use sdcHierarchies::hier_create()

add_nodes() is defunct, please use sdcHierarchies::hier_add()

delete_nodes() is defunct, please use sdcHierarchies::hier_delete()

rename_node() is defunct, please use sdcHierarchies::hier_rename()
}
\keyword{internal}
