% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dimInfo-class}
\alias{dimInfo-class}
\title{S4 class describing a dimInfo-object}
\description{
An object of class \code{dimInfo} holds all necessary information about the
dimensional variables defining a hierarchical table that needs to be protected.
}
\details{
\describe{
\item{slot \code{dimInfo}:}{a list (or NULL) with all list elements being objects of class \code{dimVar}}
\item{slot \code{strID}:}{a character vector (or NULL) defining IDs that identify each table cell. The ID's are based on (default) codes of the dimensional variables defining a cell. }
\item{slot \code{strInfo}:}{a list object (or NULL) with each list element being a numeric vector of length 2 defining the start and end-digit that is allocated by the i-th dimensional variable in ID-codes available in slot \code{strID}  }
\item{slot \code{vNames}:}{a character vector (or NULL) defining the variable names of the dimensional variables defining the table structure}
\item{slot \code{posIndex}:}{a numeric vector (or NULL) holding the position of the dimensional variables within slot \code{rawData} of class \code{dataObj} }
}
}
\note{
objects of class \code{dimInfo} are input for slots in classes \code{sdcProblem} and \code{safeObj}
}
\author{
Bernhard Meindl \email{bernhard.meindl@statistik.gv.at}
}
