% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_dimInfo.r, R/methods_class_dimInfo.r
\docType{methods}
\name{get.dimInfo}
\alias{get.dimInfo}
\alias{get.dimInfo,dimInfo,character-method}
\title{query \code{dimInfo}-objects depending on argument \code{type}}
\usage{
get.dimInfo(object, type)

\S4method{get.dimInfo}{dimInfo,character}(object, type)
}
\arguments{
\item{object}{an object of class \code{dataObj}}

\item{type}{a character vector of length 1 defining what to calculate|return|modify. Allowed types are:}
\itemize{
\item strInfo: info on how many digits in the default codes ach dimensional variable allocates
\item dimInfo: a list object with each slot containing an object of class \code{dimVar}
\item varName: variable names
\item strID: character vector of ID's defining table cells
\item posIndex vector showing the index of the elements of \code{dimInfo} in the underlying data}
}
\value{
information from objects of class \code{dimInfo} depending on argument \code{type}
\itemize{
\item a list (or NULL) if argument \code{type} matches 'strInfo', 'dimInfo'
\item numeric vector (or NULL) if argument \code{type} matches 'posIndex'
\item character vector (or NULL) if argument \code{type} matches 'varName' or 'strID'}
}
\description{
query \code{dimInfo}-objects depending on argument \code{type}
}
\note{
internal function
}
\author{
Bernhard Meindl \email{bernhard.meindl@statistik.gv.at}
}
