% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_linProb.r, R/methods_class_linProb.r
\docType{methods}
\name{get.linProb}
\alias{get.linProb}
\alias{get.linProb,linProb,character-method}
\title{query \code{linProb}-objects depending on argument \code{type}}
\usage{
get.linProb(object, type)

\S4method{get.linProb}{linProb,character}(object, type)
}
\arguments{
\item{object}{an object of class \code{linProb}}

\item{type}{a character vector of length 1 defining what to calculate|return|modify. Allowed types are:}
\itemize{
\item constraints: constraint matrix of object \code{linProb}
\item direction: directions of the constraints
\item rhs: right hand side of the constraints
\item objective: objective function
\item types: types of the objective variables
\item bounds: bounds of the objective variables}
}
\value{
information from objects of class \code{linProb} depending on \code{type}
\itemize{
\item an object of class \code{simpleTriplet} if type matches 'constraints'
\item a character vector if type matches 'direction' or 'types'
\item a numeric vector if type matches 'objective' or 'rhs'
\item a list with elements 'lower' and 'upper' if type matches 'bounds'
\itemize{
\item element 'lower': a list with the first element containing indices and the second element containing corrsponding lower bounds
\item element 'upper': a list with the first element containing indices and the second element containing corrsponding upper bounds }
}
}
\description{
query \code{linProb}-objects depending on argument \code{type}
}
\note{
internal function
}
\author{
Bernhard Meindl \email{bernhard.meindl@statistik.gv.at}
}
\keyword{internal}
