% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_dataObj.r, R/methods_class_dataObj.r
\docType{methods}
\name{init.dataObj}
\alias{init.dataObj}
\alias{init.dataObj,list-method}
\title{initialize \code{dataObj}-objects}
\usage{
init.dataObj(input)

\S4method{init.dataObj}{list}(input)
}
\arguments{
\item{input}{a list with element described below:}

\itemize{
\item element 'inputData': a list object holding data
\item element 'dimVarInd': index (within \code{inputData}) of variables that define the table to protect
\item element 'freqVarInd': index (within \code{inputData}) of variable holding frequencies
\item element 'numVarInd' index (within \code{inputData}) of numerical variables (or NULL)
\item element 'weightInd': index (within \code{inputData}) of variable holding weights (or NULL)
\item element 'sampWeightInd': index (within \code{inputData}) of variable holding sampling weights (or NULL)
\item element 'isMicroData': logical vector of length 1}
}
\value{
an object of class \code{dataObj}
}
\description{
initialize \code{dataObj}-objects
}
\note{
internal function
}
\author{
Bernhard Meindl \email{bernhard.meindl@statistik.gv.at}
}
\keyword{internal}
