% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_simpleTriplet.r,
%   R/methods_class_simpleTriplet.r
\docType{methods}
\name{init.simpleTriplet}
\alias{init.simpleTriplet}
\alias{init.simpleTriplet,character,list-method}
\title{initialize \code{simpleTriplet}-objects depending on argument \code{type}}
\usage{
init.simpleTriplet(type, input)

\S4method{init.simpleTriplet}{character,list}(type, input)
}
\arguments{
\item{type}{a character vector of length 1 defining what|how to initialize. Allowed types are:}
\itemize{
\item simpleTriplet: a simple triplet matrix
\item simpleTripletDiag: identity matrix}

\item{input}{a list depending on argument \code{type}.}
\itemize{
\item type == 'simpleTriplet': input is a list of length 1
\itemize{
\item first element: object of class 'matrix' }
\item type == 'simpleTripletDiag': input is a list of length 2
\itemize{
\item first element: numeric vector of length 1 defining the desired number of rows of the identiy matrix
\item second element: logical vector of length 1 being TRUE if a positive and FALSE if a negative identity matrix should be returned }}
}
\value{
an object of class \code{simpleTriplet}
}
\description{
init.simpleTriplet should be used to create objects of class \code{simpleTriplet}.
It is possible to create an object from class \code{simpleTriplet} from an existing matrix (using type=='simpleTriplet').
A positive (or negative) identity matrix stored as an object of class \code{simpleTriplet} can be created by specifying type=='simpleTripletDiag'.
}
\note{
internal function
}
\author{
Bernhard Meindl \email{bernhard.meindl@statistik.gv.at}
}
\keyword{internal}
