% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{microData2}
\alias{microData2}
\title{synthetic microdata}
\format{A dataframe with 100 observations on 2 variables (\code{region} and \code{gender})}
\description{
example microdata used for various examples.
}
\keyword{datasets}
