\docType{methods}
\name{print,dimVar-method}
\alias{print,dimVar-method}
\title{print \code{\link{dimVar-class}} objects}
\usage{
\S4method{print}{dimVar}(x, ...)
}
\arguments{
  \item{x}{an object used to select a method.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\description{
print \code{\link{dimVar-class}} objects in a resonable way
}

