% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_class_sdcProblem.r
\docType{methods}
\name{print,sdcProblem-method}
\alias{print,sdcProblem-method}
\title{print objects of class \code{\link{sdcProblem-class}}.}
\usage{
\S4method{print}{sdcProblem}(x, ...)
}
\arguments{
\item{x}{an objects of class \code{\link{sdcProblem-class}}}

\item{...}{currently not used.}
}
\description{
print some useful information instead of just displaying the entire object (which may be large)
}
