\docType{data}
\name{problemWithSupps}
\alias{problemWithSupps}
\title{data of class \code{\link{sdcProblem-class}}}
\format{an object of class \code{\link{sdcProblem-class}} featuring primary suppressed table cells}
\description{
  example data of class \code{\link{sdcProblem-class}} as
  created in the example of
  \code{\link{primarySuppression}}
}
\keyword{datasets}

