% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{protectedData}
\alias{protectedData}
\title{data of class \code{\link{safeObj-class}}}
\format{an object of class \code{\link{safeObj-class}} being a protected dataset}
\description{
example data of class \code{\link{safeObj-class}} as created in the example of \code{\link{protectTable}}
}
\keyword{datasets}
