% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runArgusBatchFile.R
\name{runArgusBatchFile}
\alias{runArgusBatchFile}
\title{runArgusBatchFile}
\usage{
runArgusBatchFile(obj = NULL, batchF, exe = "C:\\\\Tau\\\\TauArgus.exe")
}
\arguments{
\item{obj}{\code{NULL} or an object of class \code{\link{sdcProblem-class}} that was used to generate the batchfile for argus. If not \code{NULL},
this object is used to create correct variable names. Else, only the output from tau-Argus is read and returned as a \code{data.table}. In this case
it is possible to run tau-Argus on arbitrarily created batch-files.}

\item{batchF}{a filepath to an batch-input file created by e.g. \code{\link{createArgusInput}}.}

\item{exe}{(character) file-path to tau-argus executable}
}
\value{
a \code{data.table} containing the protected table or an error in case the batch-file was not solved correctly
}
\description{
runArgusBatchFile
}
