% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_sdcProblem.r,
%   R/methods_class_sdcProblem.r
\docType{methods}
\name{set.sdcProblem}
\alias{set.sdcProblem}
\alias{set.sdcProblem,sdcProblem,character,list-method}
\title{modify \code{sdcProblem}-objects depending on argument \code{type}}
\usage{
set.sdcProblem(object, type, input)

\S4method{set.sdcProblem}{sdcProblem,character,list}(object, type, input)
}
\arguments{
\item{object}{an object of class \code{sdcProblem}}

\item{type}{a character vector of length 1 defining what to calculate|return|modify. Allowed types are:}
\itemize{
\item problemInstance: set|modify slot 'problemInstance' of argument \code{object}
\item partition: set|modify slot 'partition' of argument \code{object}
\item startI: set|modify slot 'startI' of argument \code{object}
\item startJ: set|modify slot 'startJ' of argument \code{object}
\item indicesDealtWith: set|modify slot 'indicesDealtWith' of argument \code{object}
\item elapsedTime: set|modify slot 'elapsedTime' of argument \code{object}}

\item{input}{a list with elements depending on argument \code{type}.}

\itemize{
\item an object of class \code{problemInstance} if argument \code{type} matches 'problemInstance'
\item a list (derived from calc.multiple(type='makePartition', ...) if argument \code{type} matches 'partition'
\item a numeric vector of length 1 if argument \code{type} matches 'startI', 'startJ' or 'elapsedTime'
\item a numeric vector if argument \code{type} matches 'indicesDealtWith'}
}
\value{
an object of class \code{sdcProblem}
}
\description{
modify \code{sdcProblem}-objects depending on argument \code{type}
}
\note{
internal function
}
\author{
Bernhard Meindl \email{bernhard.meindl@statistik.gv.at}
}
