\docType{methods}
\name{show,safeObj-method}
\alias{show,safeObj-method}
\title{show \code{\link{safeObj-class}} objects}
\usage{
\S4method{show}{safeObj}(object)
}
\arguments{
  \item{object}{Any R object}
}
\description{
extract and show information stored in
\code{\link{safeObj-class}} objects
}

