% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_class_safeObj.r
\docType{methods}
\name{show,safeObj-method}
\alias{show,safeObj-method}
\title{show \code{\link{safeObj-class}} objects}
\usage{
\S4method{show}{safeObj}(object)
}
\arguments{
\item{object}{an object of class \code{\link{safeObj-class}}}
}
\description{
extract and show information stored in \code{\link{safeObj-class}} objects
}
