% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_class_sdcProblem.r
\docType{methods}
\name{show,sdcProblem-method}
\alias{show,sdcProblem-method}
\title{show objects of class \code{\link{sdcProblem-class}}.}
\usage{
\S4method{show}{sdcProblem}(object)
}
\arguments{
\item{object}{an objects of class \code{\link{sdcProblem-class}}}
}
\description{
just calls the corresponding print-method
}
