% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{simpleTriplet-class}
\alias{simpleTriplet-class}
\title{S4 class describing a simpleTriplet-object}
\description{
Objects of class \code{simpleTriplet} define matrices that are stored in a
sparse format. Only the row- and column indices and the corresponding values
of non-zero cells are stored. Additionally, the dimension of the matrix given
by the total number of rows and columns is stored.
}
\details{
\describe{
\item{slot \code{i}:}{a numeric vector specifying row-indices with each value being geq 1 and leq of the value in \code{nrRows}}
\item{slot \code{j}:}{a numeric vector specifying column-indices with each value being geq 1 and leq of the value in \code{nrCols}}
\item{slot \code{v}:}{a numeric vector specifying the values of the matrix in cells specified by the corresponding row- and column indices}
\item{slot \code{nrRows}:}{a numeric vector of length 1 holding the total number of rows of the matrix}
\item{slot \code{nrCols}:}{a numeric vector of length 1 holding the total number of columns of the matrix}
}
}
\note{
objects of class \code{simpleTriplet} are input of slot \code{constraints} in class \code{\link{linProb-class}} and slot slot \code{con} in class \code{\link{cutList-class}}
}
\author{
Bernhard Meindl \email{bernhard.meindl@statistik.gv.at}
}
