\docType{methods}
\name{summary,sdcProblem-method}
\alias{summary,safeObj-method}
\alias{summary,sdcProblem-method}
\title{summarize object of class \code{\link{sdcProblem-class}} or \code{\link{safeObj-class}}.}
\usage{
\S4method{summary}{sdcProblem}(object, ...)
}
\arguments{
  \item{object}{an object for which a summary is desired.}

  \item{...}{additional arguments affecting the summary
  produced.}
}
\description{
extract and show relevant information stored in object ofs
class \code{\link{sdcProblem-class}} or
\code{\link{safeObj-class}}.
}

