% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_class_safeObj.r
\docType{methods}
\name{summary,safeObj-method}
\alias{summary,safeObj-method}
\title{summarize \code{\link{safeObj-class}} objects}
\usage{
\S4method{summary}{safeObj}(object, ...)
}
\arguments{
\item{object}{an object of class \code{\link{safeObj-class}}}

\item{...}{additional arguments, currently ignored}
}
\description{
extract and show information stored in \code{\link{safeObj-class}} objects
}
