% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_class_sdcProblem.r
\docType{methods}
\name{summary,sdcProblem-method}
\alias{summary,sdcProblem-method}
\title{summarize object of class \code{\link{sdcProblem-class}} or \code{\link{safeObj-class}}.}
\usage{
\S4method{summary}{sdcProblem}(object, ...)
}
\arguments{
\item{object}{Objects of either class \code{\link{sdcProblem-class}} or \code{\link{safeObj-class}}.}

\item{...}{currently not used.}
}
\description{
extract and show relevant information stored in object ofs class \code{\link{sdcProblem-class}} or \code{\link{safeObj-class}}.
}
