% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeCellStatus.R
\name{changeCellStatus}
\alias{changeCellStatus}
\title{Change anonymization status of a specific cell}
\usage{
changeCellStatus(object, characteristics, varNames, rule, verbose = FALSE)
}
\arguments{
\item{object}{an object of class \linkS4class{sdcProblem}}

\item{characteristics}{a character vector specifying characteristics
of the table cell that should be identified for each dimensional variable
defining the table}

\item{varNames}{a character vector specifying variable names of dimensional
variables defining the tables}

\item{rule}{character vector of length 1 specifying a valid
anonymization code ('u', 'z', 'x', 's') to which the the cell
under consideration should be set.}

\item{verbose}{scalar logical value defining verbosity, defaults to \code{FALSE}}
}
\value{
a \linkS4class{sdcProblem} object
}
\description{
Function \code{\link[=changeCellStatus]{changeCellStatus()}} allows to change|modify the anonymization state
of single table cells for objects of class \linkS4class{sdcProblem}.
}
\note{
Important: the \code{i}-th element of argument \code{characteristics} is
used as the desired characteristic for the dimensional variable specified
at the \code{i}-th position of argument \code{varNames}!
}
\examples{
# load primary suppressed data (as created in the example
# of ?primarySuppression)
sp <- searchpaths()
fn <- paste(sp[grep("sdcTable", sp)], "/data/problemWithSupps.RData", sep="")
problem <- get(load(fn))

# we want to mark the cell region='D' and gender='male' primary sensitive
characteristics <- c('D', 'male')
varNames <- c('region', 'gender')
verbose <- TRUE
rule <- 'u'

# looking at the distribution of anonymization states before...
print(table(getInfo(problem, 'sdcStatus')))

# setting the specific cell as primary sensitive
problem <- changeCellStatus(problem, characteristics, varNames, rule, verbose)

# having a second look at the anonymization states
print(table(getInfo(problem, 'sdcStatus')))

}
\author{
Bernhard Meindl \email{bernhard.meindl@statistik.gv.at}
}
