% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createJJFormat.R
\name{createJJFormat}
\alias{createJJFormat}
\title{Create input for jj_format}
\usage{
createJJFormat(x)
}
\arguments{
\item{x}{a \linkS4class{sdcProblem} object}
}
\value{
a input suitable for \code{\link[=writeJJFormat]{writeJJFormat()}}
}
\description{
This function allows to transform a \code{sdcProblem} into a list that can
be used to  following code is made to create JJ table
format.
}
\examples{
data("microData1", package = "sdcTable")

# create hierarchies
dimList <- list(
  region = hier_create(
    root = "Total",
    nodes = LETTERS[1:4]
  ),
  gender = hier_create(
    root = "Total",
    nodes = c("male", "female")
  )
)


# creating an problem instance
prob <- makeProblem(
  data = microData1,
  dimList = dimList,
  numVarInd = "val"
)

# create inputs for jj format
inp <- createJJFormat(prob); inp

# write files to disk
# frequency table by default
writeJJFormat(inp, path = "prob_freqs.jj")

# or using the numeric variable `val` previously specified
writeJJFormat(inp, tabvar = "val", path = "prob_val.jj")
}
