% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{microdata1}
\alias{microdata1}
\title{Synthetic Microdata (1)}
\format{
a `data.frame` with `100` rows and variables `region`, `gender`
and `val`.
}
\usage{
data(microdata1)
}
\description{
A `data.frame` used for examples and problem-generation in various
examples.
}
\examples{
utils::data("microdata1", package = "sdcTable")
head(microdata1)
}
\keyword{datasets}
