% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{microdata2}
\alias{microdata2}
\title{Synthetic Microdata (2)}
\format{
a `data.frame` with `100` observations containing variables `region`,
`gender`,  `ecoOld`, `ecoNew` and `numVal`.
}
\usage{
data(microdata2)
}
\description{
Example microdata used for example in [protect_linked_tables()].
}
\examples{
utils::data("microdata2", package = "sdcTable")
head(microdata2)
}
\keyword{datasets}
