% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_class_dimVar.r
\docType{methods}
\name{print,dimVar-method}
\alias{print,dimVar-method}
\title{print \code{\link{dimVar-class}} objects}
\usage{
\S4method{print}{dimVar}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{dimVar-class}}}

\item{...}{currently not used}
}
\description{
print \code{\link{dimVar-class}} objects in a resonable way
}
