\docType{data}
\name{problem}
\alias{problem}
\title{data of class \code{\link{sdcProblem-class}}}
\format{an object of class \code{\link{sdcProblem-class}}}
\description{
  example data of class \code{\link{sdcProblem-class}} as
  created in the example of \code{\link{makeProblem}}
}
\keyword{datasets}

