% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.r
\name{runArgusBatchFile}
\alias{runArgusBatchFile}
\title{runArgusBatchFile}
\usage{
runArgusBatchFile(obj, batchF, exe = "C:\\\\Tau\\\\TauArgus.exe")
}
\arguments{
\item{obj}{an object of class \code{\link{sdcProblem-class}} that was used to generate the batchfile for argus.
This is used to create corrent variable names.}

\item{batchF}{a filepath to an batch-input file crated by \code{\link{createArgusInput}}}

\item{exe}{(character) file-path to tau-argus executable}
}
\value{
a \code{data.table} with the protected table
}
\description{
runArgusBatchFile
}
