% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.r
\name{sdcProb2df}
\alias{sdcProb2df}
\title{sdcProb2df}
\usage{
sdcProb2df(obj, addDups = TRUE, addNumVars = FALSE,
  dimCodes = "both")
}
\arguments{
\item{obj}{an object of class \code{\link{sdcProblem-class}} from \code{sdcTable}}

\item{addDups}{(logical), if \code{TRUE}, duplicated cells are included in the output}

\item{addNumVars}{(logical), if \code{TRUE}, numerical variables (if defined in \code{\link{makeProblem}}) will
be included in the output.}

\item{dimCodes}{(character) allows to specify in which coding the dimensional variables should be returned. Possible choices are:
\itemize{
\item "both": both original and internally used, standardized codes are included in the output
\item "original": only original codes of dimensional variables are included in the output
\item "default": only internally used, standardized codes are included in the output
}}
}
\value{
a \code{data.table} containing information about all cells of the given sdc problem
instance is returned.
}
\description{
returns a \code{data.frame} or \code{data.table} from an \code{\link{sdcProblem-class}}-object which contains the
current state of the problem instance.
}
\examples{
## have a look at ?makeProblem
}
