\docType{methods}
\name{summary,safeObj-method}
\alias{summary,safeObj-method}
\alias{summary,sdcProblem-method}
\title{summarize \code{\link{safeObj-class}} objects}
\usage{
\S4method{summary}{safeObj}(object, ...)

\S4method{summary}{sdcProblem}(object, ...)
}
\arguments{
  \item{object}{an object for which a summary is desired.}

  \item{...}{additional arguments affecting the summary
  produced.}
}
\description{
extract and show information stored in
\code{\link{safeObj-class}} objects

extract and show relevant information stored in
\code{\link{sdcProblem-class}} objects
}

